Crate typeck

Structs

Identifier
A part of an IR that contains an identifier.

Mutability
A part of a local or property whichi contains information about it's mutability. Properties use var while locals use let mut.

Ty
A type, the meat of the sandwhich. Ty represents a type which is used to represent a specific type. Type info is generated or inferred by context. A Ty can be inferred or generated depending upon building blocks or sister componenets. Smart casting using the given context to ensure that while within a conditional block that checks for a type's instance, that we safely cast an object's type to the checked type.

TyValue
A value from input and it's type. This can include primitive data such as literals, Unit values (aka, void or nothing), or custom types when parsing Constructors. See TyValueElement for more information.

Typeck
A single instance of a type checker, thus the shortened name Typeck. Each file is given its own Typeck. Explicit source module declarations are part of the IR and are not given their own typecks. Example:

TypeckManager
A global manager for all [typeck]s. All typeck's are added to a threadpool upon a call to [enqueueModule]. A single notice send channel is shared between all typeck's. The flow of notices currently is temporary but looks something like this

Enums

TyValueElement
An element of a TyValue. This has to do with primitive data since primitives are built in. This also allows for convenience when checking for non-primitives types. If you have a class called A, then the TyValueElement for A would be `TyValueElement::Custom("A"). This will then be used during type checking to ensure that Values such as calling A's constructor is matched with the required type of the statement.

Traits

GetTy
A trait that provides a method called get_ty which is a convenience method for quickly getting an IR element's type info.

Load
This trait provides an associated function for loading typeck IR into the current typeck instance. Output is what type is being returned upon success. Due to the fact that traits don't have known sizes at compiletime, an associated type will do. This trait is implemented for different kinds of IR such as a [statements::Statement] or [properties::Property].